Renamed to ninchat titlebar
